requirements.txt
flask==2.3.2
flask-sqlalchemy==3.0.3
python-dotenv==1.0.0
flask-login==0.6.2
flask-migrate==4.0.4
psycopg2-binary==2.9.6
flask-wtf==1.1.1
email-validator==2.0.0.post2
flask-admin==1.6.1

